requirements.txt
streamlit
openai
pdfplumber
python-dotenv
langchain-text-splitters
langchain-community
chromadb
dashscope